crates_io
Struct NewCrate

pub struct NewCrate {

    pub name: String,
    pub vers: String,
    pub deps: Vec<NewCrateDependency>,
    pub features: BTreeMap<String, Vec<String>>,
    pub authors: Vec<String>,
    pub description: Option<String>,
    pub documentation: Option<String>,
    pub homepage: Option<String>,
    pub readme: Option<String>,
    pub readme_file: Option<String>,
    pub keywords: Vec<String>,
    pub categories: Vec<String>,
    pub license: Option<String>,
    pub license_file: Option<String>,
    pub repository: Option<String>,
    pub badges: BTreeMap<String, BTreeMap<String, String>>,
    pub links: Option<String>,
    pub rust_version: Option<String>,

}

This struct is serialized as JSON and sent as metadata ahead of the crate tarball when publishing crates to a crate registry like crates.io.

see https://doc.rust-lang.org/cargo/reference/registry-web-api.html#publish

Fields
name: Stringvers: Stringdeps: Vec<NewCrateDependency>features: BTreeMap<String, Vec<String>>authors: Vec<String>description: Option<String>documentation: Option<String>homepage: Option<String>readme: Option<String>readme_file: Option<String>keywords: Vec<String>categories: Vec<String>license: Option<String>license_file: Option<String>repository: Option<String>badges: BTreeMap<String, BTreeMap<String, String>>links: Option<String>rust_version: Option<String>